% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersupport.R
\name{BFGS_lpdf}
\alias{BFGS_lpdf}
\title{Do optimization}
\usage{
BFGS_lpdf(om, logpdf, rho = list(), newt = F, ...)
}
\arguments{
\item{om}{An initial point}

\item{logpdf}{An initial point}

\item{rho}{An initial point}

\item{...}{additional parameters passed to \code{obBFGS}}
}
\value{
A list of information from optimization
}
\description{
Do optimization
}
