% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersupport.R
\name{BFGS_std}
\alias{BFGS_std}
\title{BFGS standard}
\usage{
BFGS_std(funcw, rho, ..., verbose = 0)
}
\arguments{
\item{funcw}{An object to optimize}

\item{rho}{An initial point}

\item{...}{additional parameters passed to \code{funcw}}

\item{verbose}{Integer from 0-3 where larger prints more information}
}
\value{
A list of information from optimization
}
\description{
Do generic minimization of a function \code{funcw} that takes
a list rho using the Broyden-Fletcher-Goldfarb-Shanno algorithm.
Useful for hyperparameter optimization because it handles infs 
fairly easily.
}
