% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covf}
\alias{covf}
\alias{Rcpp_covf-class}
\alias{Rcpp_covf}
\title{covariance function class}
\description{
This is a base class designed to handle the specific features of 
covariances needed for outerbase.  Polymorphism allows for the implied 
methods to be used across several similar classes.
}
\section{Fields}{

\describe{
\item{\code{covf$hyp}}{hyperparameters for this specific correlation function}

\item{\code{covf$lowbnd,covf$uppbnd}}{upper and lower bounds for the inputs to the 
covariance function.}

\item{\code{covf$cov(x1,x2)}}{returns the covariance matrix between two vectors of 
inputs \code{x1} and \code{x2}}

\item{\code{covf$covdiag(x1)}}{returns the diagonal of the covariance matrix between 
\code{x1} and itself}

\item{\code{covf$cov_gradhyp(x1,x2)}}{returns a cube of the gradient the \code{cov} 
with respect to the covariance hyperparameters}
}}

\seealso{
derived class: \code{\link{covf_mat25}}, 
\code{\link{covf_mat25pow}}, 
\code{\link{covf_mat25ang}}
}
