% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covf_mat25}
\alias{covf_mat25}
\alias{Rcpp_covf_mat25-class}
\alias{Rcpp_covf_mat25}
\title{Matern covariance function}
\description{
\preformatted{
covf = new(covf_mat25)
}
This is the standard Matern covariance function which has form
\deqn{c(x_1,x_2) = (1+ |h| + h^2/3) \exp(-|h|) }
where \eqn{h = (x_1-x_2)/\rho} and \eqn{\rho}=\code{exp(2*hyp[0])}.
}
\seealso{
base class: \code{\link{covf}}
}
