% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covf_mat25pow}
\alias{covf_mat25pow}
\alias{Rcpp_covf_mat25pow-class}
\alias{Rcpp_covf_mat25pow}
\title{Matern covariance function with power transform}
\description{
\preformatted{
covf = new(covf_mat25pow)
}
This is the standard Matern covariance function with a power transformation
which has form
\deqn{c(x_1,x_2) = (1+ |h| + h^2/3) \exp(-|h|) }
where \eqn{h = (x_1^\alpha-x_2^\alpha)/\rho} and \code{hyp} is a two 
dimensional vector with \eqn{\rho}=\code{exp(2*hyp[0]+0.25*hyp[1])}
and \eqn{\alpha}=\code{exp(2*hyp[0])}.
}
\seealso{
base class: \code{\link{covf}}
}
