% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitob}
\alias{fitob}
\title{fit an outerbase}
\usage{
fitob(x, y, numb = 100, covnames = NULL)
}
\arguments{
\item{x}{a n by d sized matrix of inputs}

\item{y}{a n length vector of outputs}

\item{numb}{number of bases to use}

\item{covnames}{a d length vector of covariance names}
}
\value{
Saving important model information
}
\description{
fit an outerbase
}
