% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gethyp}
\alias{gethyp}
\title{Get the hyperparameters}
\arguments{
\item{om}{\code{\link{outermod}} object}
}
\description{
\preformatted{
hyp = gethyp(om)
}
This function gets the current hyperparameters from an outermod object. It 
formats them in a way that makes reading in `R` helpful.
}
\examples{
om = new(outermod)
setcovfs(om, c("mat25", "mat25", "mat25"))
hyp = gethyp(om)
print(hyp)
}
\seealso{
\code{\link{outermod}}
}
