% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getpara}
\alias{getpara}
\title{Get the model parameters}
\arguments{
\item{lpdf}{logpdf object}
}
\description{
\preformatted{
para = getpara(loglik)
}
This function gets the current parameters from an \code{\link{lpdf}} object.  
It formats them in a way that makes reading in `R` helpful.
}
