% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{listcov}
\alias{listcov}
\title{list all covariance functions}
\usage{
listcov()
}
\value{
list all names of covariance functions recommend
as of this edition.  The first is the default.
}
\description{
list all covariance functions
}
