% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_gda}
\alias{loglik_gda}
\alias{Rcpp_loglik_gda-class}
\alias{Rcpp_loglik_gda}
\title{Gaussian errors with diagonal adjustment}
\arguments{
\item{om}{an \code{\link{outermod}} object to be referred to}

\item{terms}{a matrix of \code{terms}, must have as many columns as dims in 
\code{om}}

\item{y}{a vector of observations}

\item{x}{a matrix of predictors, must have as many columns as dims in 
\code{om} and the same number of rows as \code{y}}
}
\description{
\preformatted{
loglik = new(loglik_gda, om, terms, y, x)
}
This is a standard model which has the form
\deqn{y = \langle \phi(x), \theta \rangle + \delta(x) + \varepsilon,
\delta(x) \sim N(0, \lambda g(x)), \varepsilon \sim N(0,\sigma^2)}
where \eqn{\phi(x)} is the basis, \eqn{\theta} is the coefficient vector,
\eqn{\delta(x)} is unseen vector corresponding to unmodeled 
variance \eqn{\lambda g(x)}, \eqn{\varepsilon} is an unseen noise vector.
The parameter vector is of length 2 where 
\eqn{\sigma=} \code{exp(para[0])} and \eqn{\lambda=}\code{exp(2 para[1])}.
}
\seealso{
base class: \code{\link{lpdf}}
}
