% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logpr_gauss}
\alias{logpr_gauss}
\alias{Rcpp_logpr_gauss-class}
\alias{Rcpp_logpr_gauss}
\title{Gaussian prior}
\arguments{
\item{om}{an \code{\link{outermod}} object to be referred to}

\item{terms}{a matrix of \code{terms}, must have as many columns as dims in 
\code{om}}
}
\description{
\preformatted{
logpr = new(logpr_gauss, om, terms)
}
This is a standard model of coefficents which has them as drawn independently
from
\deqn{ \theta_i \sim N(0, \rho c_i)}
where \eqn{c_i} is the variance supplied by \code{om} for the $i$th term. 
The parameter vector is of length 1 where 
\eqn{\rho=} \code{exp(para[0])}.
}
\seealso{
base class: \code{\link{lpdf}}
}
