% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf$optcg}
\alias{lpdf$optcg}
\title{Optimization via Conjugate Gradient}
\arguments{
\item{tol}{A positive double representing tolerance, default is 
\code{0.001}.}

\item{epoch}{A positive integer representing the maximum number of steps 
conjugate gradient will take.}
}
\description{
\preformatted{
lpdf$optcg(tol,epoch)
}
This optimizes the coefficient vector \code{coeff} using conjugate gradient. 
It currently is designed only for quadratic \code{\link{lpdf}} objects.
}
\seealso{
\code{\link{lpdf}}
}
