% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdf$optnewton}
\alias{lpdf$optnewton}
\title{Optimization via Newton's Method}
\description{
\preformatted{
lpdf$optnewton()
}
This optimizes the coefficient vector \code{coeff} using Newton's Method.  
It currently is designed only for quadratic \code{\link{lpdf}} objects.  
It should take a single step.
}
\seealso{
\code{\link{lpdf}}
}
