% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersupport.R
\name{lpdfwrapper}
\alias{lpdfwrapper}
\title{Optimization wrapper}
\usage{
lpdfwrapper(hypl, om, logpdf, newt = F)
}
\arguments{
\item{hypl}{parameter list}

\item{om}{an \code{\link{outermod}} object}

\item{logpdf}{a \code{\link{lpdf}} object}

\item{newt}{bool for if newtons method should be used}
}
\value{
list of \code{val} and \code{grad}
}
\description{
Optimization wrapper
}
