% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{obpred}
\alias{obpred}
\title{Prediction from outerbase}
\usage{
obpred(obmodel, x)
}
\arguments{
\item{obmodel}{output from \code{\link{obfit}}}

\item{x}{a new m by d sized matrix of inputs}
}
\value{
A list with \code{mean} and \code{var} at new x
}
\description{
This function allows for turning an \code{obmodel} into predictions with 
mean and variance.
}
\seealso{
obfit
}
