% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfuncs.R
\name{obtest_borehole3d}
\alias{obtest_borehole3d}
\title{Three dim borehole example}
\usage{
obtest_borehole3d(x)
}
\arguments{
\item{x}{a n by 3 vector of inputs}
}
\value{
a length n vector of outputs
}
\description{
A three dimensional Borehole function used in illustrations.
}
