% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfuncs.R
\name{obtest_borehole8d}
\alias{obtest_borehole8d}
\title{Eight demensional Borehole}
\usage{
obtest_borehole8d(x)
}
\arguments{
\item{x}{a n by 8 vector of inputs}
}
\value{
a length n vector of outputs
}
\description{
Eight demensional Borehole
}
