% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outerbase$build}
\alias{outerbase$build}
\title{Builds the outerbase}
\description{
\preformatted{
outerbase$build()
 }
Build (or re-build) a basis based on the recent evaluation 
of \code{\link{outermod}}.
}
\seealso{
\code{\link{outerbase}}
}
