% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outerbase$getbase}
\alias{outerbase$getbase}
\title{Get base functions}
\arguments{
\item{k}{An integer from that corresponds to the dimension.}
}
\value{
basis_func A matrix of evaluated basis functions for that 
dimension.  Designed mostly for visualization.
}
\description{
\preformatted{
basis_func = outerbase$getbase(k)
}
Returns the basis for a dimension
}
\seealso{
\code{\link{outerbase}}
}
