% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outerbase$matmul}
\alias{outerbase$matmul}
\title{Matrix multiply}
\arguments{
\item{terms}{A matrix of \code{terms}}

\item{a}{A vector of length the same as the rows in \code{terms}}
}
\value{
b A vector resulting from the matrix multiplication
}
\description{
\preformatted{
 b = outerbase$matmul(terms, a)
}
Multiplies the basis times a vector without building the basis 
matrix.
}
\seealso{
\code{\link{outerbase}}
}
