% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outerbase$tmatmul}
\alias{outerbase$tmatmul}
\title{Transpose Matrix multiply}
\arguments{
\item{terms}{A matrix of \code{terms}}

\item{a}{A vector of length the same as the rows in \code{outerbase}}
}
\value{
b A vector resulting from the matrix multiplication
}
\description{
\preformatted{
  b = outerbase$tmatmul(terms, a)
}
Multiplies the transpose of the basis times a vector without 
 building the basis matrix.
}
\seealso{
\code{\link{outerbase}}
}
