% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersupport.R
\docType{package}
\name{outerbase-package}
\alias{outerbase-package}
\alias{_PACKAGE}
\title{outerbase}
\description{
High-dimensional function approximation using outer product models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mattplumlee.github.io/outerbase/index.html}
  \item \url{https://github.com/MattPlumlee/outerbase}
  \item Report bugs at \url{https://github.com/MattPlumlee/outerbase/issues}
}

}
\author{
\strong{Maintainer}: Matthew Plumlee \email{mplumlee@northwestern.edu}

}
