% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outermod$getvar}
\alias{outermod$getvar}
\title{Get variance of coefficients}
\arguments{
\item{terms}{A matrix of \code{terms}}
}
\value{
coeffvar A vector of variances of each coefficient
}
\description{
\preformatted{
coeffvar = outermod$getvar(terms)
}
Returns the variance of the coefficients associated with \code{terms}.
}
\seealso{
\code{\link{outermod}}
}
