% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outermod$selectterms}
\alias{outermod$selectterms}
\title{Select optimal terms}
\arguments{
\item{numterms}{Number of basis \code{terms} desired}
}
\value{
terms A matrix of \code{terms}
}
\description{
\preformatted{
terms = om$selectterms(numterms)
}
Selects the best \code{terms} given the current \code{outermod}
}
\seealso{
\code{\link{outermod}}
}
