% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outermod}
\alias{outermod}
\alias{Rcpp_outermod-class}
\alias{Rcpp_outermod}
\title{Outer product-type model}
\description{
Type the name of the class to see its methods.
}
\section{Fields}{

\describe{
\item{\code{\link{setcovfs}}}{to set covariance functions}

\item{\code{\link{setknot}}}{to set knot points}

\item{\code{\link{gethyp}}}{to get hyperparameters}

\item{\code{\link{outermod$updatehyp}}}{to update hyperparameters}

\item{\code{\link{outermod$selectterms}}}{to find best terms}

\item{\code{\link{outermod$getvar}}}{to find variances of coefficients}
}}

\examples{
om = new(outermod)
setcovfs(om, c("mat25", "mat25", "mat25"))
setknot(om,
         list(seq(0,1,by=0.01),seq(0,1,by=0.01),seq(0,1,by=0.01)))
terms = om$selectterms(40)
coeffvar =om$getvar(terms)
hyp = gethyp(om)
hyp[1:2] = 0.5
om$updatehyp(hyp)
coeffvar = om$getvar(terms)
}
\seealso{
\code{\link{outerbase}} the main product from outermods
}
