% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictor}
\alias{predictor}
\alias{Rcpp_predictor-class}
\alias{Rcpp_predictor}
\title{prediction class}
\arguments{
\item{logpdf}{An \code{\link{lpdf}} instance to build the predictor}
}
\description{
\preformatted{
pred = new(predictor, logpdf)
}
This is a base class design to allow for coherent building of
predictions across multiple models.  Unlike many base classes in this 
package, it is meant to be directly used.
}
\section{Fields}{

\describe{
\item{\code{predictor$update(x)}}{update the current input for prediction}

\item{\code{predictor$mean()}}{return the current mean of the prediction}

\item{\code{predictor$var()}}{return the current var of the prediction}
}}

