% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{predob}
\alias{predob}
\title{pred from an outerbase}
\usage{
predob(obmodel, x)
}
\arguments{
\item{obmodel}{output from \code{\link{fitob}}}

\item{x}{a new m by d sized matrix of inputs}
}
\value{
A list with \code{mean} and \code{var} at new x
}
\description{
pred from an outerbase
}
