% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setcovfs}
\alias{setcovfs}
\title{Set covariance functions}
\arguments{
\item{om}{\code{\link{outermod}} object}

\item{covnames}{vector of strings of the covariance functions}
}
\description{
\preformatted{
setcovfs(om, covnames)
}
This function sets the covariance functions for an outermod object,
which is first thing one does when creating an outermod object.//'
}
\examples{
om = new(outermod)
setcovfs(om, c("mat25", "mat25", "mat25"))
setcovfs(om, c("mat25", "mat25pow", "mat25", "mat25ang"))
}
\seealso{
\code{\link{outermod}}
}
