% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setknot}
\alias{setknot}
\title{Set knot points}
\arguments{
\item{om}{\code{\link{outermod}} object}

\item{knotslist}{list of one dimensional vectors}
}
\description{
\preformatted{
setknot(om, knotslist)
}
This function sets the knot points to estimate the eigenfunctions
and eigenvalues. It will naturally check if the knot points have the
same dimension as the covariance functions.  It will also check if the 
knot points are within reasonable bounds for the covariance functions.//'
}
\examples{
om = new(outermod)
setcovfs(om, c("mat25", "mat25", "mat25"))
setknot(om,
         list(seq(0,1,by=0.01),seq(0,1,by=0.01),seq(0,1,by=0.01)))
}
\seealso{
\code{\link{outermod}}, \code{\link{setcovfs}}
}
