% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersupport.R
\name{BFGS_lpdf}
\alias{BFGS_lpdf}
\title{BFGS lpdf}
\usage{
BFGS_lpdf(om, logpdf, rho = list(), newt = F, ...)
}
\arguments{
\item{om}{an \code{\link{outermod}} object}

\item{logpdf}{a \code{\link{lpdf}} object}

\item{rho}{an initial point, initialized from objects if needed}

\item{newt}{bool for if Newtons method should be used}

\item{...}{additional parameters passed to \code{\link{BFGS_std}}}
}
\value{
A list of information from optimization
}
\description{
A wrapper for code{\link{BFGS_std}} that is useful for easily calling 
parameter optimization for this package with as few lines as possible.
Note that \code{om} and \code{logpdf} will be set to optimal 
parameters, the return is simply for information.
}
