% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersupport.R
\name{BFGS_std}
\alias{BFGS_std}
\title{Do optimization}
\usage{
BFGS_std(funcw, rho, ..., Bs = NULL, lr0 = 0.1, verbose = 0)
}
\arguments{
\item{funcw}{An object to optimize}

\item{rho}{An initial point}

\item{...}{additional parameters passed to \code{ll}}

\item{Bs}{initial inverse hessian}

\item{lr0}{initial learning rate}

\item{verbose}{Integer from 0-3 where larger prints more information}
}
\value{
A list of information from optimization
}
\description{
Do optimization
}
