% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getpara}
\alias{getpara}
\title{Get the model parameters}
\arguments{
\item{logpdf}{an \code{\link{lpdf}} class instance}
}
\value{
a vector of parameters
}
\description{
\preformatted{
para = getpara(logpdf)
}
This function gets the current parameters from an \code{\link{lpdf}} class
instance. It formats them in a way that makes reading in \code{R} easier.
}
