% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpdfvec}
\alias{lpdfvec}
\alias{Rcpp_lpdfvec-class}
\alias{Rcpp_lpdfvec}
\title{Vector of \code{lpdf} instances}
\arguments{
\item{loglik}{one reference to a \code{lpdf} instance}

\item{logpr}{another reference to a \code{lpdf} instance that shares 
\code{\link{outermod}}  with \code{loglik}}
}
\value{
no returns, this is a class which contains methods
}
\description{
\preformatted{
logpdf = new(lpdfvec, loglik, logpr)
}
This is a class where each instance contains two \code{\link{lpdf}} 
instances and can be 
manipulated as a single instance.  It presumes both are based on the same
\code{\link{outermod}} instance, thus they share hyperparameters.  However
the model parameters are concatenated.  Currently also includes variations
on marginal adjustments.  

Currently it is designed only for a pair, but the ordering is arbitrary.
}
\section{Fields}{

\describe{
\item{\code{lpdfvec$domarg}}{A boolean that controls if marginal adjustment is 
done}
}}

\seealso{
base class: \code{\link{lpdf}}
}
