% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outerbase$getbase}
\alias{outerbase$getbase}
\title{Get base functions}
\arguments{
\item{k}{An integer from that corresponds to the dimension.}
}
\value{
a matrix of evaluated basis functions
}
\description{
\preformatted{
basis_func = outerbase$getbase(k)
}
Returns the basis for dimension \code{k}.   Designed mostly for 
visualization.
}
\seealso{
\code{\link{outerbase}}
}
