% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outerbase$getmat}
\alias{outerbase$getmat}
\title{Get basis matrix}
\arguments{
\item{terms}{a matrix of terms}
}
\value{
a matrix of evaluated basis functions based on 
\code{terms}.
}
\description{
\preformatted{
basismat = outerbase$getmat(terms)
 }
Returns the basis matrix for a given set of \code{terms}.
}
\seealso{
\code{\link{outerbase}}
}
