% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersupport.R
\docType{package}
\name{outerbase-package}
\alias{outerbase-package}
\alias{_PACKAGE}
\title{outerbase}
\description{
High-dimensional regression using outer product models. Research on the methods is currently under investigation and published resources will be posted as they are available. As the method is new, the website is the best resource for understanding the principals. Some of the core ideas are based on Plumlee and coauthors' work on analysis of grid-structured experiments described in Plumlee (2014) \doi{10.1080/01621459.2014.900250} and Plumlee, Erickson, Ankenman, Lawrence (2021) \doi{10.1093/biomet/asaa084}. Some additional textbooks for additional information on Gaussian processes are Rasmussen and Williams (2005) \doi{10.7551/mitpress/3206.001.0001} and Gramacy (2022) \doi{10.1201/9780367815493}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mattplumlee.github.io/outerbase/}
  \item \url{https://github.com/MattPlumlee/outerbase/}
  \item Report bugs at \url{https://github.com/MattPlumlee/outerbase/issues}
}

}
\author{
\strong{Maintainer}: Matthew Plumlee \email{mplumlee@northwestern.edu}

}
