% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outermod$selectterms}
\alias{outermod$selectterms}
\title{Select optimal terms}
\arguments{
\item{numterms}{number of basis \code{terms} desired}
}
\value{
a matrix of \code{terms}
}
\description{
\preformatted{
terms = om$selectterms(numterms)
}
Returns the best \code{numterms} given \code{outermod} currently using
maximum variance criteria.
}
\seealso{
\code{\link{outermod}}
}
