% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outermod$updatehyp}
\alias{outermod$updatehyp}
\title{Update hyperparameters}
\arguments{
\item{hyp}{A vector of hyperparameters}
}
\value{
no value is returned, the class instance is updated
}
\description{
\preformatted{
outermod$updatehyp(hyp)
}
Updates the hyperparameters for the instance of outermod.
}
\seealso{
\code{\link{outermod}}
}
