% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{outermod}
\alias{outermod}
\alias{Rcpp_outermod-class}
\alias{Rcpp_outermod}
\title{Outer product-type model}
\value{
no returns, this is a class which contains methods
}
\description{
This is a class used to construct \code{\link{outerbase}}
class instances.  It stores key information for constructing a basis.
}
\section{Fields}{

\describe{
\item{\code{\link{outermod$updatehyp}(hyp)}}{update hyperparameters}

\item{\code{\link{outermod$selectterms}(numterms)}}{find best \code{numterms} terms}

\item{\code{\link{outermod$getvar}(terms)}}{find variances of coefficients 
associated with  \code{terms}}
}}

\examples{
om = new(outermod)
setcovfs(om, c("mat25", "mat25", "mat25"))
setknot(om,
         list(seq(0,1,by=0.01),seq(0,1,by=0.01),seq(0,1,by=0.01)))
terms = om$selectterms(40)
coeffvar =om$getvar(terms)
hyp = gethyp(om)
hyp[1:2] = 0.5
om$updatehyp(hyp)
coeffvar = om$getvar(terms)
}
\seealso{
\code{\link{outerbase}} the main product from an outermod

\code{\link{setcovfs}}, \code{\link{setknot}}, 
\code{\link{gethyp}}
}
