% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictor}
\alias{predictor}
\alias{Rcpp_predictor-class}
\alias{Rcpp_predictor}
\title{prediction class}
\arguments{
\item{loglik}{An \code{\link{lpdf}} instance, specifically that starts with 
\code{loglik}, to build the predictor}
}
\value{
no returns, this is a class which contains methods
}
\description{
\preformatted{
pred = new(predictor, loglik)
}
This is a base class design to allow for coherent building of
predictions across multiple models.  Unlike many base classes in this 
package, it is meant to be directly used.
}
\section{Fields}{

\describe{
\item{\code{predictor$update(x)}}{update the current input to \code{x} for prediction}

\item{\code{predictor$mean()}}{return the vector of means for the prediction}

\item{\code{predictor$var()}}{return the vector of variances for the prediction}

\item{\code{predictor$setnthreads(k)}}{specifics \code{k} as the number of threads
to use}
}}

