% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setcovfs}
\alias{setcovfs}
\title{Set covariance functions}
\arguments{
\item{om}{an \code{\link{outermod}} instance}

\item{covnames}{a vector of strings of the covariance functions}
}
\value{
no value is returned, \code{om} is updated
}
\description{
\preformatted{
setcovfs(om, covnames)
}
Sets the covariance functions for an outermod class instance.
This is first thing one does when creating an outermod instance.
}
\examples{
om = new(outermod)
setcovfs(om, c("mat25", "mat25", "mat25"))
setcovfs(om, c("mat25", "mat25pow", "mat25", "mat25ang"))
}
\seealso{
\code{\link{outermod}}
}
